DE-3484 Refactor utils.py into two separate files under a utils folder. This is to fix a cyclic dependency when check_for_mutual_ambiguity is in the common utils file
