+++ __tests__/unit/test_ts_datetime.py
@@ -2,7 +2,7 @@ import pytest
 from pendulum import now
 
 from task_script_utils.datetime_parser import ts_datetime
-from task_script_utils.datetime_parser.utils import from_pendulum_format
+from task_script_utils.datetime_parser.utils.conversions import from_pendulum_format
 
 
 subseconds_test_cases = [
