+++ task_script_utils/convert_datetime_to_ts_format.py
@@ -1,5 +1,5 @@
 from typing import Optional
-import dateparser, arrow, datetime
+import dateparser, arrow
 from dateutil import tz
